Parallel Queue Bug resolve

+++ AuxPort/Core/Utility/AuxUtility.cpp
@@ -154,9 +154,9 @@ std::string AuxPort::Utility::join(const std::vector<std::string>& tokens, char 
 std::string AuxPort::Utility::convertToTime(uint32_t seconds)
 {
     auto secondsInFloat = static_cast<float>(seconds);
-    float hours = floorf(secondsInFloat / 3600);
-    float mins = floorf((secondsInFloat - (hours * 3600)) / 60);
-    float secs = floorf(secondsInFloat - (hours * 3600) - (mins * 60));
+    uint32_t hours = static_cast<uint32_t>(floorf(secondsInFloat / 3600));
+    uint32_t mins = static_cast<uint32_t>(floorf((secondsInFloat - (hours * 3600)) / 60));
+    uint32_t secs = static_cast<uint32_t>(floorf(secondsInFloat - (hours * 3600) - (mins * 60)));
     return std::to_string(hours) + " : " + std::to_string(mins) + " : " + std::to_string(secs);
 }
